requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy<2
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0 